pyproject.toml
[project]
name = "ssh-file-explorer"
version = "0.1.0"
description = "SSH/CLI File Explorer with fuzzy search and audio playback, and image preview"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "textual>=3.1.0",
    "pillow>=11.2.1",
    "rich-pixels>=3.0.1",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.1",
    "pdfminer-six>=20250416",
]
